chore: rename picked type file name
